refactor(message_list): only chat user messages will be shown
Prior to this change bot and user messages where shown in the
wh00t chat client.  This changes the bot messages to not show
up by default, unless the client is in DEBUG mode.  This will
most likely be in place until a sane option for how to handle
bot messages on the wh00t chat client is developed.

Closes #3